inference/server_requirements.txt
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
huggingface_hub
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart
sentencepiece
protobuf
einops
psutil
gpustat
pyyaml
python-dotenv